+++ connectors/routing/connector_test.go
@@ -24,6 +24,7 @@ import (
 	"context"
 	"reflect"
 	"sort"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -49,6 +50,17 @@ var (
 			buildRouter("default", "default", "memory"),
 		},
 	}
+	routersStr = strings.Join([]string{
+		"{production.map -> memory}",
+		"{production.default -> random}",
+		"{ebook.ebook-store -> memory}",
+		"{ebook.default -> random}",
+		"{ebook.apple.* -> memory}",
+		"{ebook.* -> devnull}",
+		"{development.map -> memory}",
+		"{development.default -> random}",
+		"{default.default -> memory}",
+	}, ",")
 	testInfo = &dosa.EntityInfo{
 		Ref: &dosa.SchemaRef{
 			Scope:      "production",
@@ -170,6 +182,7 @@ func TestGetConnector(t *testing.T) {
 	// no plugin
 	// glob match
 	rc := NewConnector(cfg, connectorMap)
+	assert.Equal(t, "[Routing ["+routersStr+"]]", rc.String())
 	ei := &dosa.EntityInfo{
 		Ref: &dosa.SchemaRef{Scope: "ebook", NamePrefix: "apple.v1"},
 	}